fix(basin): use Token auth scheme, register plugin, unskip endpoint tests
