+++ packages/corsair/core/constants.ts
@@ -55,6 +55,7 @@ export const BaseProviders = [
 	'ayrshare',
 	'basecamp',
 	'baselinker',
+	'basin',
 	'betterstack',
 	'bigmailer',
 	'bigml',
@@ -222,6 +223,7 @@ export const ProviderDisplayNames = {
 	ayrshare: 'Ayrshare',
 	basecamp: 'Basecamp',
 	baselinker: 'BaseLinker',
+	basin: 'Basin',
 	betterstack: 'Better Stack',
 	bigmailer: 'BigMailer',
 	bigml: 'BigML',
@@ -396,6 +398,7 @@ export type AllProviders =
 	| 'ayrshare'
 	| 'basecamp'
 	| 'baselinker'
+	| 'basin'
 	| 'betterstack'
 	| 'bigmailer'
 	| 'bigml'
